% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_helper.R
\name{break_month}
\alias{break_month}
\title{break month}
\usage{
break_month(input, block = 6, time_type = "month")
}
\arguments{
\item{time_type}{}
}
\description{
break month
}
\author{
Dongqiang Zeng
}
