% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tme_classifier.R
\name{tme_classifier}
\alias{tme_classifier}
\title{TME classifier for gastric cancer}
\usage{
tme_classifier(
  eset,
  method = "ensemble",
  tme_data = NULL,
  tme_deconvolution = FALSE,
  array = FALSE,
  save_data = FALSE,
  adjust_eset = FALSE,
  replace_na = TRUE,
  log2trans = FALSE,
  scale = FALSE,
  perm = 100
)
}
\arguments{
\item{eset}{expression set with genes names as row and sample id as column}

\item{method}{default is `ensemble`, other options: `svm`, `rf`, `nnet`, `knn`, `dt`, `xgboost`}

\item{tme_data}{user can provide TME data estimated by CIBERSORT and MCPcounter, please modify the names of cell types, a reference data could be found by `eset_example`}

\item{tme_deconvolution}{default is FALSE, if set to TRUE, this process will be time consuming.}

\item{array}{default is FALSE, if expression set was derived from micro array, this parameter should be set to TRUE, which will affect the TME deconvolution process}

\item{save_data}{default is FALSE, if TRUE, processing data will be save to local path}

\item{adjust_eset}{default is FALSE, it means that variables with missing value will to be preserved and will be replaced by mean value}

\item{replace_na}{default is TRUE, variables with missing value will to be preserved and will be replaced by mean value of observations}

\item{log2trans}{default is FALSE,}

\item{scale}{default is FALSE}

\item{perm}{permutation of CIBERSORT}
}
\description{
TME classifier for gastric cancer
}
\examples{
tme_classifier(eset = eset_example1, method = "ensemble", scale = T)
}
\author{
Dongqiang Zeng
}
