% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tme_classifier.R
\name{tme_classifier}
\alias{tme_classifier}
\title{TME classifier for gastric cancer}
\usage{
tme_classifier(
  eset,
  method = "ensemble",
  tme_data = NULL,
  tme_deconvolution = FALSE,
  array = FALSE,
  save_data = FALSE,
  adjust_eset = FALSE,
  replace_na = TRUE,
  log2trans = FALSE,
  scale = TRUE,
  perm = 100,
  slot = NULL,
  assays = NULL,
  group.by = "orig.ident",
  source_gene_length = "local",
  min_prob = NULL
)
}
\arguments{
\item{eset}{expression set with genes names as row and sample id as column, or Seurat object}

\item{method}{default is `ensemble`, other options: `svm`, `rf`, `nnet`, `knn`, `dt`, `xgboost`}

\item{tme_data}{user can provide TME data estimated by CIBERSORT and MCPcounter, please modify the names of cell types, a reference data could be found by `eset_example`}

\item{tme_deconvolution}{default is FALSE, if set to TRUE, this process will be time consuming.}

\item{array}{default is FALSE, if expression set was derived from micro array, this parameter should be set to TRUE, which will affect the TME deconvolution process}

\item{save_data}{default is FALSE, if TRUE, processing data will be save to local path}

\item{adjust_eset}{default is FALSE, it means that variables with missing value will to be preserved and will be replaced by mean value}

\item{replace_na}{default is TRUE, variables with missing value will to be preserved and will be replaced by mean value of observations}

\item{log2trans}{default is FALSE,}

\item{scale}{default is TRUE}

\item{perm}{permutation of CIBERSORT}

\item{slot}{if input is 'Seurat' object, slot should be defined to aggregate scRNA matrix group by samples; options: `counts`, `SCT`, `integrated`}

\item{assays}{default is null. If null, DefaultAssay will be used}

\item{group.by}{default is `orig.ident`}

\item{source_gene_length}{default is `web`, which will request data from the `Ensembl` database, if the error is returned,try `default`.}

\item{min_prob}{minimal probability for the subtype}
}
\description{
TME classifier for gastric cancer
}
\examples{
tme_classifier(eset = eset_example1, method = "ensemble", scale = T)
}
\author{
Dongqiang Zeng
}
